All that I am able to do
Here is the final merge as I am ale to do it. It is incomplete as I
cannot get the buttons to work:
My problem is that I cannot pass an associative array from PHP to a
script function. Using serialise/unserialise should work, but it does
not perform as expected...